email/requirements.txt
pydantic==2.8.0
python-dotenv==1.0.0
# python_version >= "3.11"
